ANDROID: sched/core: Add a latency-sensitive flag to uclamp
Add a 'latency_sensitive' flag to uclamp in order to express the need
for some tasks to find a CPU where they can wake-up quickly. This is not
expected to be used without cgroup support, so add solely a cgroup
interface for it.

As this flag represents a boolean attribute and not an amount of
resources to be shared, it is not clear what the delegation logic should
be. As such, it is kept simple: every new cgroup starts with
latency_sensitive set to false, regardless of the parent.

In essence, this is similar to SchedTune's prefer-idle flag which was
used in android-4.19 and prior.

Change-Id: I722d8ecabb428bb7b95a5b54bc70a87f182dde2a
Signed-off-by: Quentin Perret <quentin.perret@arm.com>
(cherry picked from commit ae6a2a277c8f9bf9dcedc1ecbc3e61dabc930cbb)
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
(cherry picked from commit 08e153fda576b98ccd048dd8462e4239d94f8e1b)
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>