treewide: Add unified libcam cmdline parameters
libcameraservice.so is dropped by a developer to fix video recording on Redmi Note 10, dropping libcameraservice.so requires dropping [1] aswell. Dropping libcameraservice.so and [1] fixed video recording on Android 12 but broke usage of ANX. This however causes a fragmentation, hence I came up with the idea to unify [1] with the help of the userland_worker.

[1]: 680de38f70de {"msm: camera: Use boot clock for recording start time"}

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>